QPID-3603: c++ messaging API: allow floating point reconnect durations in seconds.
Allow sub-second intervals, e.g. reconnect_interval_min=0.001 for a
millisecond interval.

git-svn-id: https://svn.apache.org/repos/asf/qpid/branches/qpid-3603-2@1243579 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/client/amqp0_10/ConnectionImpl.h
@@ -64,10 +64,10 @@ class ConnectionImpl : public qpid::messaging::ConnectionImpl
     std::vector<std::string> urls;
     qpid::client::ConnectionSettings settings;
     bool reconnect;
-    int64_t timeout;
+    double timeout;
     int32_t limit;
-    int64_t minReconnectInterval;
-    int64_t maxReconnectInterval;
+    double minReconnectInterval;
+    double maxReconnectInterval;
     int32_t retries;
     bool reconnectOnLimitExceeded;
